requirements.txt
moviepy==1.0.3
Pillow==9.0.0
requests==2.27.1
requests_oauthlib==1.3.0
